requirements.txt
torch>=2.0
torchvision>=0.15
pillow
numpy
opencv-python
scikit-learn
pytorch-grad-cam
skyfield
